+++ hapi-fhir-base/src/main/resources/META-INF/services/ca.uhn.fhir.repository.impl.IRepositoryLoader
@@ -1,1 +1,1 @@
-ca.uhn.fhir.repository.impl.InMemoryFhirRepositoryLoader
+ca.uhn.fhir.repository.impl.memory.InMemoryFhirRepositoryLoader
